+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## 3.18.9
+
+- Added `rejected.testing.AsyncTestCase.published_messages` and `rejected.testing.PublishedMessage`
+- Updated testing documentation
+- Updated the setup.py extras install for testing to install all testing dependencies
+- Made `raven` optional in `rejected.testing`
+
 ## 3.18.8
 
 - Fix the mocks in `rejected.testing`